dash/parsers: create parseAdaptationSets function
